refactor(xod-deploy, xod-client-electron): move library manager from `xod-arduino-deploy` into `xod-deploy`
