+++ packages/xod-arduino-deploy/src/index.js
@@ -5,9 +5,3 @@ export { installArchitecture, installTool } from './packageManager';
 export { listPorts, openPort, openAndReadPort, closePort } from './serialport';
 export { default as packageIndex } from './packageIndex.json';
 export { listBoardsFromIndex, parseFQBN, strigifyFQBN } from './utils';
-export { default as messages } from './messages';
-export {
-  getLibraryNameFromUrl,
-  checkLibrariesInstalledByUrls,
-  installLibrariesByUrls,
-} from './libraryManager';
