--- packages/xod-deploy/test/mocha.opts
@@ -1,2 +1,0 @@
---require babel-register
---colors
